- Added InoicePay - Added InvoiceRequest - Added Qr code for InvoicePay - Added json validator - Added Invoice  to the Spotlight
