+++ app/assets/stylesheets/components/_all.scss
@@ -30,6 +30,7 @@
 @import "direct-debit";
 @import "collapsible_table";
 @import "voting";
+@import "invoice";
 
 .lifetime {
     background: url("../fresh-bolt2.png") no-repeat 100% 50%;
