+++ app/assets/stylesheets/components/_invoice.scss
@@ -1,0 +1,8 @@
+.invoice-request-input {
+    .ant-col.ant-form-item-label {
+        text-align: left !important;
+    }
+    :after {
+        content: none !important;
+    }
+}
