+++ package.json
@@ -163,6 +163,7 @@
         "intro.js": "^2.8.0-alpha.1",
         "jdenticon": "git+https://github.com/cryptonomex/jdenticon.git",
         "js-sha256": "^0.2.3",
+        "jsonschema": "^1.2.4",
         "jspdf": "^1.4.1",
         "jspdf-autotable": "^3.1.1",
         "lodash-es": "^4.17.10",
